Cleanup logging handling
