+++ crates/cli/src/lib.rs
@@ -1,7 +1,7 @@
 #![deny(rust_2018_idioms)]
 #![allow(clippy::missing_const_for_fn, clippy::future_not_send)]
 
-use std::{env::var, fs::File, io::Write, process::Stdio, sync::Mutex};
+use std::{io::Write, process::Stdio};
 
 use args::{Args, ShellCompletion};
 use clap::CommandFactory;
@@ -9,8 +9,8 @@ use clap_complete::{Generator, Shell};
 use clap_mangen::Man;
 use is_terminal::IsTerminal;
 use miette::{IntoDiagnostic, Result};
-use tokio::{fs::metadata, io::AsyncWriteExt, process::Command};
-use tracing::{debug, info, warn};
+use tokio::{io::AsyncWriteExt, process::Command};
+use tracing::{debug, info};
 use watchexec::Watchexec;
 use watchexec_events::{Event, Priority};
 
@@ -22,82 +22,6 @@ mod dirs;
 mod emits;
 mod filterer;
 mod state;
-
-async fn init() -> Result<Args> {
-	let mut log_on = false;
-
-	#[cfg(feature = "dev-console")]
-	match console_subscriber::try_init() {
-		Ok(_) => {
-			warn!("dev-console enabled");
-			log_on = true;
-		}
-		Err(e) => {
-			eprintln!("Failed to initialise tokio console, falling back to normal logging\n{e}")
-		}
-	}
-
-	if !log_on && var("RUST_LOG").is_ok() {
-		match tracing_subscriber::fmt::try_init() {
-			Ok(()) => {
-				warn!(RUST_LOG=%var("RUST_LOG").unwrap(), "logging configured from RUST_LOG");
-				log_on = true;
-			}
-			Err(e) => eprintln!("Failed to initialise logging with RUST_LOG, falling back\n{e}"),
-		}
-	}
-
-	let args = args::get_args().await?;
-	let verbosity = args.verbose.unwrap_or(0);
-
-	if log_on {
-		warn!("ignoring logging options from args");
-	} else if verbosity > 0 {
-		let log_file = if let Some(file) = &args.log_file {
-			let is_dir = metadata(&file).await.map_or(false, |info| info.is_dir());
-			let path = if is_dir {
-				let filename = format!(
-					"watchexec.{}.log",
-					chrono::Utc::now().format("%Y-%m-%dT%H-%M-%SZ")
-				);
-				file.join(filename)
-			} else {
-				file.to_owned()
-			};
-
-			// TODO: use tracing-appender instead
-			Some(File::create(path).into_diagnostic()?)
-		} else {
-			None
-		};
-
-		let mut builder = tracing_subscriber::fmt().with_env_filter(match verbosity {
-			0 => unreachable!("checked by if earlier"),
-			1 => "warn",
-			2 => "info",
-			3 => "debug",
-			_ => "trace",
-		});
-
-		if verbosity > 2 {
-			use tracing_subscriber::fmt::format::FmtSpan;
-			builder = builder.with_span_events(FmtSpan::NEW | FmtSpan::CLOSE);
-		}
-
-		match if let Some(writer) = log_file {
-			builder.json().with_writer(Mutex::new(writer)).try_init()
-		} else if verbosity > 3 {
-			builder.pretty().try_init()
-		} else {
-			builder.try_init()
-		} {
-			Ok(()) => info!("logging initialised"),
-			Err(e) => eprintln!("Failed to initialise logging, continuing with none\n{e}"),
-		}
-	}
-
-	Ok(args)
-}
 
 async fn run_watchexec(args: Args) -> Result<()> {
 	info!(version=%env!("CARGO_PKG_VERSION"), "constructing Watchexec from CLI");
@@ -192,8 +116,7 @@ async fn run_completions(shell: ShellCompletion) -> Result<()> {
 }
 
 pub async fn run() -> Result<()> {
-	let args = init().await?;
-	debug!(?args, "arguments");
+	let args = args::get_args().await?;
 
 	if args.manual {
 		run_manpage(args).await
